feature(src): реализовано добавление резистора на канвас
